requirement.txt
Package        Version
-------------- --------
ConfigUpdater  3.1.1
numpy          1.26.0
opencv-python  4.8.0.76
packaging      23.1
Pillow         10.0.1
pip            23.2.1
platformdirs   3.10.0
pycairo        1.24.0
PyGObject      3.46.0
setuptools     68.0.0
setuptools-scm 8.0.3
SimpleCV       1.3
tomlkit        0.12.1
wheel          0.38.4
